Converting array into object in parseClassOutput.
The class output syntax follows the ( "," result ) convention where result is ( variable = value ). This is closer to an object rather than an array. The complication comes if there are duplicate variable names in the result. What we do in that case is turn the "variable" entry in the object into an array, and add to that the duplicate entries.

Not sure if this is the best approach, but having an object is better than an array when possible (easier indexing/search).